fix: Googleログイン設定、supabase設定
