+++ app/auth/callback/route.ts
@@ -1,5 +1,5 @@
 import { NextResponse } from "next/server";
-import { createSupabaseActionClient } from "@/lib/supabase/server-action";
+import { createSupabaseActionClient } from "@/lib/supabase/supabase-server";
 
 export async function GET(request: Request) {
   const url = new URL(request.url);
