+++ app/companies/actions.ts
@@ -2,7 +2,7 @@
 
 import { revalidatePath } from "next/cache";
 import { redirect } from "next/navigation";
-import { createSupabaseActionClient } from "@/lib/supabase/server-action";
+import { createSupabaseActionClient } from "@/lib/supabase/supabase-server";
 
 export async function createCompany(formData: FormData) {
   const supabase = await createSupabaseActionClient();
