+++ app/dashboard/_components/sign-out-button.tsx
@@ -3,7 +3,7 @@
 import { useRouter } from "next/navigation";
 import { useState } from "react";
 
-import { createSupabaseBrowserClient } from "@/lib/supabase/client";
+import { createSupabaseBrowserClient } from "@/lib/supabase/supabase-browser";
 
 export default function SignOutButton() {
   const [loading, setLoading] = useState(false);
