--- lib/supabase/server.ts
@@ -1,2 +1,0 @@
-// Default to readonly client to avoid cookie writes in Server Components.
-export { createSupabaseReadonlyClient as createSupabaseServerClient } from "./server-readonly";
