+++ lib/supabase/supabase-browser.ts
@@ -4,10 +4,8 @@ import { Database } from "../database.types";
 export const createSupabaseBrowserClient = () => {
   const supabaseUrl = process.env.NEXT_PUBLIC_SUPABASE_URL;
   const supabaseAnonKey = process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY;
-
   if (!supabaseUrl || !supabaseAnonKey) {
     throw new Error("Supabase URL/Anon Key is missing. Check env.");
   }
-
   return createBrowserClient<Database>(supabaseUrl, supabaseAnonKey);
 };
